refactor: improve tool data management and remove unused storage
- Move backup/restore methods to McpEditorWindow for better encapsulation
- Extract ConnectedLLMToolData to separate file for reusability
- Remove ConnectedLLMToolsStorage as it's no longer needed
- Simplify McpServerController by delegating tool management to McpEditorWindow
- Clean up unnecessary imports and migration code
