+++ Packages/src/Editor/Core/ConnectedLLMToolData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5789a9d996dc84bb68c28099fde785c5
+guid: a4320ff5004aa40e9ae756b620cc25a1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
